+++ types/context.go
@@ -30,7 +30,9 @@ type Context struct {
 }
 
 // create a new context
-func NewContext(ms MultiStore, header abci.Header, isCheckTx bool, txBytes []byte, logger log.Logger) Context {
+func NewContext(ms MultiStore, header abci.Header, isCheckTx bool,
+	txBytes []byte, logger log.Logger, absentValidators []int32) Context {
+
 	c := Context{
 		Context: context.Background(),
 		pst:     newThePast(),
@@ -43,6 +45,7 @@ func NewContext(ms MultiStore, header abci.Header, isCheckTx bool, txBytes []byt
 	c = c.WithIsCheckTx(isCheckTx)
 	c = c.WithTxBytes(txBytes)
 	c = c.WithLogger(logger)
+	c = c.WithAbsentValidators(absentValidators)
 	return c
 }